Fetch all the conversation IDs from an Intercom App
To fetch all the Conversations from Intercom, there are a couple of
things that we need to do, as well as some funky new stuff we're going
to be doing.

const scrapeIntercom = async () => {
  let idList = await fetchAllConversationIds();
  console.log(JSON.stringify(idList, null, 4));
}

scrapeIntercom();

Firstly, we're going to use async/await. The async keyword indicates
that at some point within this function, we're going to be waiting on a
promise from somewhere else. The await keyword will wait for that
Promise to be resolved, before assiging that value to a variable and
using it. In this case, fetching all the conversation ids from intercom
will take time and many, many fetch requests, which is why we're using
async/await.

If we step down into the fetchAllConversationIds function, we see it
assembling an array of function calls. The INTERCOM_PAGE_COUNT is
hardcoded for now (we'll fix it later).

for (i = 1; i <= INTERCOM_PAGE_COUNT; i++) {
  let promiseThrottle = new PromiseThrottle({
    requestsPerSecond: PROMISE_THROTTLE_RPS,
    promiseImplementation: Promise
  });
  allConversations.push(promiseThrottle.add(fetchConversationList.bind(this, i)));
}

We're using a library called `PromiseThrottle`, simply so the promises
don't all fire at once and DDOS the Intercom API, but runs 5 at a time.
This is slow, but is probably better than Intercom invalidating our
access token. We simply push a binded function call into this array,
but we don't actually call it yet. In this case, we're binding the
`fetchConversationList` method with a page number. The condition of
using `Promise.all`, as well as `PromiseThrottle`, is that these
functions have to return promises. `fetch` does this (return a promise),
as it makes heavy useage of promises itself.

return Promise.all(allConversations)
  .then(function(idList) {
    return idList;
  }).catch((err) => console.error(err));

Here, we run all the promises in our array, and once they're all done,
it moves onto the `then` method. `Promise.all` will then concatenate
these results into an array itself. The `return idList` resolves the
entire `Promise.all`, so back in the `scrapeIntercom` function, the
`await` is resolved, and the value assigned.

Ta-dah! How to do concurrent promises, 101.